Backend:Team <DONE>

+++ app/Http/Controllers/Site/TeamController.php
@@ -8,25 +8,46 @@ use App\Team;
 
 class TeamController extends Controller
 {
+    public function __construct() 
+    {
+        $this->middleware('auth');
+    }
+
     public function index() 
     {
+        if (auth()->user()->level !== "ADMIN") {
+            return redirect()->back()->with('error', 'Unauthorized Page');
+        }
+
     	$teams = Team::paginate(15);
     	return view('site.team.index', compact("teams"));
     }
 
     public function show($id) 
     {
+        if (auth()->user()->level !== "ADMIN") {
+            return redirect()->back()->with('error', 'Unauthorized Page');
+        }
+
     	$teams = Team::findOrFail($id);
     	return view('site.team.show', compact("teams"));
     }
 
     public function create() 
     {
+        if (auth()->user()->level !== "ADMIN") {
+            return redirect()->back()->with('error', 'Unauthorized Page');
+        }
+
     	return view('site.team.create');
     }
 
     public function store(Request $request) 
     {
+        if (auth()->user()->level !== "ADMIN") {
+            return redirect()->back()->with('error', 'Unauthorized Page');
+        }
+
     	$msg = [
     		"required" => "Form tidak boleh kosong",
     		"max" => "Karakter maksimal 12"
@@ -56,12 +77,20 @@ class TeamController extends Controller
 
     public function edit($id) 
     {   
+        if (auth()->user()->level !== "ADMIN") {
+            return redirect()->back()->with('error', 'Unauthorized Page');
+        }
+
         $team = Team::findOrFail($id);
         return view('site.team.edit', compact("team"));
     }
 
     public function update(Request $request, $id) 
     {
+        if (auth()->user()->level !== "ADMIN") {
+            return redirect()->back()->with('error', 'Unauthorized Page');
+        }
+
         $msg = [
             "required" => "Form tidak boleh kosong",
             "max" => "Karakter maksimal 12"
@@ -91,9 +120,13 @@ class TeamController extends Controller
 
     public function destroy($id) 
     {
+        if (auth()->user()->level !== "ADMIN") {
+            return redirect()->back()->with('error', 'Unauthorized Page');
+        }
+
         $team = Team::findOrFail($id);
         $team->delete();
 
-        return redirect()->back()->with('success', 'Data Berhasil Dihapus!');
+        return redirect()->back()->with('success', 'Data successfully deleted!');
     }
 }
